pyproject.toml
[project]
name = "ugc-video-analysis"
version = "0.1.0"
description = "Slack-integrated UGC video review system using Gemini AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "slack-bolt>=1.18.0",
    "aiohttp>=3.9.0",
    "google-genai>=0.4.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
